consistency_check/requirements.txt
connexion[flask,uvicorn,swagger-ui]
httpx
pyyaml